std::map operator[] creates an entry if one is not found. Thus, we do not need to explicitly initialize the map.
This ensures that we only create entries in the map
if that type is used in any tracking operation.

Reduces the number of types we think need metadata.
Might help reduce the number of bits needed in the 
metadata, if we can find a upper limit in some
large examples.


git-svn-id: https://llvm.org/svn/llvm-project/poolalloc/trunk@131900 91177308-0d34-0410-b5e6-96231b3b80d8