Overdue and simple fix for infinite recursion bug in BU.
When deciding to recalculate the DSGraph due to "new knowledge"
(new resolvable callsites introduced while inlining), we only
do so if the new callsites add new callee Functions that weren't
previously being inlined before.

This does mean we'll miss some precision, and to track how often this fires I've
introduced a new statistic that counts when this happens.  Running on CINT2006
this fires exactly once (with almost no effect on the result), so I'm calling it
well worth the potential loss of precision (which seems extraordinarily minor)
for the sake of not going into infinite recursion on some codes.

Note that the unresolved callsites we skip processing will be resolved in the
callers of the current function, so the loss is localized to this particular
graph.

Update the previously XFAIL'd lit test capturing this, it now passes!  Hooray!

For posterity's sake (and perhaps future-Will's sake), I'll add the following:

Ideally this check wouldn't exist, and instead we could recalculate whenever new
callsites were added.  "New" being an interestingly nebulous thing when
callsites are implicitly named by their acyclic call paths, something we do not
explicitly track.  If we did, we should use that to compare if the new callsites
are duplicates of what we had before, but the way we do things now a comparison
of DSCallsites does not capture the call paths used to reach them.

I've sat on this change for quite a while because of this theoretical
deficiency, but I'm tired of DSA crashing on programs due to this and having
measured the resulting precision difference this deviation from 'correct'
doesn't seem to matter.  Hopefully this only a positive effect for DSA and all
of its users.

Onward! :)

git-svn-id: https://llvm.org/svn/llvm-project/poolalloc/trunk@159124 91177308-0d34-0410-b5e6-96231b3b80d8