Add test for BU,TD behavior with indirect-indirect chains.
XFAIL for now, due to (intentional?) behavior where
TD reports the same callgraph found during BU.

This means calls that are resolved in caller
(FP passed in as argument, for example) will
not be reported as resolved in the callgraph.

Investigation is needed to sort out why TD
reports BU's callgraph instead of computing
and reporting a callgraph of its own.

Modifying TD to compute its own callgraph
causes this test to pass, FWIW.



git-svn-id: https://llvm.org/svn/llvm-project/poolalloc/trunk@238489 91177308-0d34-0410-b5e6-96231b3b80d8